refactoring of SummaryDocumentationCodeFixProvider
